fix: change time type Int to Double

+++ data/src/main/kotlin/team/duckie/app/android/data/quiz/model/GetQuizResponse.kt
@@ -22,7 +22,7 @@ internal data class GetQuizResponse(
     @JsonProperty("score")
     val score: Int? = null,
     @JsonProperty("time")
-    val time: Int? = null,
+    val time: Double? = null,
     @JsonProperty("user")
     val user: UserResponse? = null,
     @JsonProperty("wrongProblem")
