+++ data/src/main/kotlin/team/duckie/app/android/data/quiz/repository/QuizRepositoryImpl.kt
@@ -28,7 +28,7 @@ class QuizRepositoryImpl @Inject constructor(
     override suspend fun submitQuiz(
         examId: Int,
         correctProblemCount: Int,
-        time: Int?,
+        time: Double?,
         requirementAnswer: String?,
         problemId: Int?,
         wrongAnswer: String?,
